Move Form components into directories
