+++ assets/javascripts/kitten/components/form/checkbox/checkbox.stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { text, boolean } from '@storybook/addon-knobs'
 import { action } from '@storybook/addon-actions'
 import { Checkbox } from './checkbox'
-import { Grid, GridCol } from '../../components/grid/grid'
+import { Grid, GridCol } from '../../../components/grid/grid'
 
 export default {
   component: Checkbox,
