+++ assets/javascripts/kitten/components/form/comment-avatar/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { CommentAvatar as DefaultCommentAvatar } from '../../components/comments/comment-avatar'
-import { Deprecated } from '../../helpers/utils/deprecated'
+import { CommentAvatar as DefaultCommentAvatar } from '../../../components/comments/comment-avatar'
+import { Deprecated } from '../../../helpers/utils/deprecated'
 
 export const CommentAvatar = props => {
   const message = `You can use the CommentAvatar instead.
