+++ assets/javascripts/kitten/components/form/comment-form/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { CommentForm as DefaultCommentForm } from '../../components/comments/comment-form'
-import { Deprecated } from '../../helpers/utils/deprecated'
+import { CommentForm as DefaultCommentForm } from '../../../components/comments/comment-form'
+import { Deprecated } from '../../../helpers/utils/deprecated'
 
 export const CommentForm = props => {
   const message = `You can use the CommentForm instead.
