+++ assets/javascripts/kitten/components/form/form-amount-and-currency/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { TextInputAndSelect } from '../../components/form/text-input-and-select'
+import { TextInputAndSelect } from '../../../components/form/text-input-and-select'
 
 export const FormAmountAndCurrency = props => {
   const {
