+++ assets/javascripts/kitten/components/form/rich-text-input/index.js
