+++ assets/javascripts/kitten/components/form/switch/index.js
